+++ tests/collaborative/ot/ot_sheet_deleted.test.ts
@@ -21,7 +21,6 @@ import {
   RenameSheetCommand,
   ResizeColumnsRowsCommand,
   SetBorderCommand,
-  SetDecimalCommand,
   SetFormattingCommand,
   ShowSheetCommand,
   UpdateCellCommand,
@@ -102,11 +101,6 @@ describe("OT with DELETE_SHEET", () => {
     row: 0,
     border: undefined,
   };
-  const setDecimal: Omit<SetDecimalCommand, "sheetId"> = {
-    type: "SET_DECIMAL",
-    target: [toZone("A1")],
-    step: 3,
-  };
   const createChart: Omit<CreateChartCommand, "sheetId"> = {
     type: "CREATE_CHART",
     id: "1",
@@ -160,7 +154,6 @@ describe("OT with DELETE_SHEET", () => {
     setFormatting,
     clearFormatting,
     setBorder,
-    setDecimal,
     createChart,
     resizeColumns,
     resizeRows,
